feat(KToaster): switch to setup syntax [KHCP-6522]
