+++ src/types/index.ts
@@ -26,3 +26,4 @@ export * from './prompt'
 export * from './pop'
 export * from './skeleton'
 export * from './select'
+export * from './toaster'
